backend/requirements-updated.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP client for Ollama API
httpx>=0.25.0

# RAG and AI
langchain>=0.1.0
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Document processing - Multiple options for better compatibility
# Primary PDF processing
pymupdf>=1.24.0

# Alternative PDF processing if PyMuPDF fails
pdfplumber>=0.9.0

# Word document processing
python-docx>=1.1.0

# Additional document processing
pdfminer.six>=20221105
unstructured[pdf,docx]>=0.11.0

# Additional utilities
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
matplotlib>=3.7.0

# MCP (Model Context Protocol)
mcp>=1.0.0 